yaw.deprecated now mirrors the structure of yaw
